Add strings to Settings.Secure for notification permission
The Settings.Secure string is used for the "review notification permissions" notification that will be shown to users upon upgrade to T, and will track state for whether the user has seen and interacted with the notification yet.

The action is an internal intent to go to the app notification settings page. This intent and activity need to be non-exported to disallow other apps from causing writes to settings.

Bug: 225373531
Test: manual with change to show the notification
Change-Id: I3ad1a3043eb428fab93e15d8f5ae1ada884b010f