[DO NOT MERGE] Force extra nav bar provide side gesture insets
This is a patch to fix the missing gesture insets issue when there's a
task bar and no navigation bar on the display. It will try to add the
gesture navigation insets when a window providing navigation bar insets
is added and there's no window with TYPE_NAVIGATION_BAR.

The logic will only be applied when the insets is not flexible. Due to
the removal of the old hard-coded logic in the later version, this patch
should not be merged into the next version.

Bug: 212643340
Test: android.widget.cts.SeekBarTest#testSetOnSeekBarChangeListener
Change-Id: I080df3547361aef1aa259dd2b3dc9bdb860a0193
(cherry picked from commit 7b0571cb163ae4775134c1a1bae5faaa538d5a7f)
Merged-In: I080df3547361aef1aa259dd2b3dc9bdb860a0193