Remove ViewDebug calls from ActivityThread
This change replaces calls to ViewDebug that are intended to count
instances of View/ViewRootImpl by adding them to the list of classes
counted by VMDebug.countInstancesOfClasses. This is functionally a no-op
change because getViewInstanceCount and getViewRootImplCount both call
VMDebug.countInstancesOfClass.

This results in a ~4.5% reduction in CPU cycles during a bug report (and
a ~19% reduction in VMDebug_countInstancesOfClass[es]).

Test: Build; check that the HIGH meminfo bug report section still
      contains the same set of object counts.
Bug: 215574756
Change-Id: I3f70c30cbf61f8e57fdf67b681d73f89b3027d72