Fix notifications accepting touches outside their bounds
I have confirmed that this bug is the result of the ENR's mClipTopAmount being negative, but I am not certain why; my best guess is that the launch animation is causing this value to be set to a negative value, and that it's not cleaning up at the end of the animation.  However, that alone isn't enough to explain why this can keep happening on subsequent taps, so more investigation is needed there.  On the other hand, this is the safest way to fix the tap issue without risking the introduction of other bugs.

Test: manual testing using breakpoint evaluation to apply the logic change while in the bad state
Fixes: 230080345
Change-Id: I3dec0e99111777ca37eddf79a33487fc0c8dc99c