Updates face-based auto-rotate triggering conditions.
Auto-rotate with face detection wasn't enabled when the keyguard is
locked. This affects Maps navigation or Google Meets that require
auto-rotate even the keyguard is locked. This change updates the
condition so that the feature will disabled only when the keyguard is
showing on the lock screen.

Test: atest WindowOrientationListenerTest
Test: Manually tested with Google Meet
Bug: 226375593

Change-Id: Icbf8981f5606ba8b82e1d1d4006a2017fbe30f0f
(cherry picked from commit bcdfae60c1ad5acba0388c0e823b27f616055866)