Ignore group alert setting when opening bubbles
When notifications are shown in a group, group alert setting can be set
to GROUP_ALERT_SUMMARY to notify the user only once per group.
Regardless of this alert setting, user is still allowed to bubble
individiual notifications in a group (button is shown).
Issue was that our check if a notification can be bubbled was applying
the check for alerting on a child notification in a group. Which is not
needed as bubbling should be allowed regardless.
Removing the check for bubbles.

Bug: 214130620
Test: atest NotificationInterruptStateProviderImplTest
Test: manually verified that when there are multiple notifications from
Telegram, they can be opened as bubbles
Test: manually verified with Messages that notifications in a group can
be opened as bubbles

Change-Id: I01258e0db041b8eb04160242f4fd9a875315ad3d