Run unlock animation when device is already unlocked
When the screen times out, the device may not immediately lock based
upon a user setting. In this case, keyguard is visible but not fully
locked. If the user uses finger auth, MODE_ONLY_WAKE is used but the
keyguard is not properly dismissed. Use MODE_WAKE_AND_UNLOCK to run
all animations.

Fixes: 231553219
Test: atest BiometricsUnlockControllerTest
Change-Id: I8ff2c2f983b64c7f62d7025fce227a2dd9eabf04