Position bubbles from right to left
When a RTL language is used, position bubbles from right to left when
the stack is expanded.
BubblePositioner uses view index to position bubbles on screen.
Reverting the index in case of RTL languages so calculations are done
with last bubble in the list landing on the left side etc.

Bug: 234071989
Test: manual, on phone, set device language to Arabic, expand bubble stack,
observe overflow icon on the left, top bubble on the right
Test: manual, on tablet, set device language to Arabic, expand bubble
stack, obserer top bubble on top, overflow at the bottom

Change-Id: I297f1e653046d488e6de9ac81eb6bfb8d3afef1d