Fix a bug in maximize buttons in legacy captions
The bug was introduced in 44ee85262863e19e8a0fc53cc9ddbaa2c74c35b6 which
effectively removed the bounds resolution for fullscreen tasks.
Therefore we need to remove the bounds manually.

Also hide caption on shell behind the flag. Note we still let WM shell
handle freeform resizing so that we don't have to reintroduce an
exception for freeform windows to WM core.

Bug: 165794636
Test: Maximize buttons in legacy captions work.
Test: Caption in WM shell isn't shown when the feature is disabled and
shown when it's enabled.

Change-Id: I51d15556dabdc6713f31d7515c574dfb0db49b7a