Forward device updates when info changes
In some cases, the device/route name could change without also updating
the ID. This updates the check to include all fields of MediaDeviceData except
for the icon (which can change reference more frequently than is necessary
to reload). Now if the name, status, or intent has changed, listeners
will be updated to have the new info.

Bug: 235627592
Test: manual
Test: atest MediaDataManagerTest
Change-Id: Iefd7ce0e8859b89db2bc75c243b3a04a0cb3915d